--- cirkit/layers/factorized/__init__.py
@@ -1,2 +1,0 @@
-from .cp import CPLayer as CPLayer
-from .sum_product import SumProductLayer as SumProductLayer
