+++ cirkit/layers/sum_product/__init__.py
@@ -1,0 +1,2 @@
+from .base import SumProductLayer as SumProductLayer
+from .cp import CPLayer as CPLayer
