+++ cirkit/layers/sum_product/base.py
