+++ cirkit/models/__init__.py
@@ -1,1 +1,1 @@
-from .pcs import TensorizedPC as TensorizedPC
+from .tensorized_circuit import TensorizedPC as TensorizedPC
