--- cirkit/models/pcs/__init__.py
@@ -1,1 +1,0 @@
-from .tensorized_pc import TensorizedPC as TensorizedPC
